+++ V9_IberTrade1.0/tfg/src/main/java/org/tfg/spring/tfg/controller/ZapatillaController.java
@@ -14,12 +14,14 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
+import org.tfg.spring.tfg.domain.Usuario; // Asegúrate de tener importada esta clase
 import org.tfg.spring.tfg.exception.DangerException;
 import org.tfg.spring.tfg.helper.PRG;
 import org.tfg.spring.tfg.service.MarcaService;
 import org.tfg.spring.tfg.service.ModeloService;
 import org.tfg.spring.tfg.service.ZapatillaService;
 
+import jakarta.servlet.http.HttpSession;
 
 @RequestMapping("/zapatilla")
 @Controller
@@ -37,18 +39,23 @@ public class ZapatillaController {
     private static final String LOAD_DIR = "src\\main\\resources\\static\\img";
 
     @GetMapping("r")
-    public String r(
-            ModelMap m,@Param("PalabraClave")String palabraClave) {
+    public String r(ModelMap m, @Param("PalabraClave") String palabraClave, HttpSession session) {
+        Usuario usuario = (Usuario) session.getAttribute("usuario");
+        if (usuario == null || !Boolean.TRUE.equals(usuario.getAdmin())) {
+            return "redirect:/login";
+        }
         m.put("zapatillas", zapatillaService.findAll(palabraClave));
-        m.put("palabraClave",palabraClave);
+        m.put("palabraClave", palabraClave);
         m.put("view", "zapatilla/r");
         return "_t/frame";
     }
 
     @GetMapping("c")
-    public String c(
-            ModelMap m) {
-    
+    public String c(ModelMap m, HttpSession session) {
+        Usuario usuario = (Usuario) session.getAttribute("usuario");
+        if (usuario == null || !Boolean.TRUE.equals(usuario.getAdmin())) {
+            return "redirect:/login";
+        }
         m.put("marcas", marcaService.findAll());
         m.put("modelos", modeloService.findAll());
         m.put("view", "zapatilla/c");
@@ -56,38 +63,42 @@ public class ZapatillaController {
     }
 
     @PostMapping("c")
-    public String cPost(
-            @RequestParam("nombre") String nombre,
-            @RequestParam("precio") double precio,
-            @RequestParam("color") String color,
-            @RequestParam("talla") String talla,
-            @RequestParam("stock") Integer stock,
-            @RequestParam("idMarca") Long idMarca,
-            @RequestParam("idModelo") Long idModelo,
-            @RequestParam("imagenZapatilla") MultipartFile imagenZapatilla
-            ) throws DangerException
-             {if(!imagenZapatilla.isEmpty()){
-        try {
-            
-            String imageName = imagenZapatilla.getOriginalFilename();
-            Path destPath = Paths.get(LOAD_DIR, imageName);
-            Files.copy(imagenZapatilla.getInputStream(), destPath, StandardCopyOption.REPLACE_EXISTING);
-            String nombreImagen = imageName.substring(imageName.lastIndexOf("\\") + 1, imageName.lastIndexOf("."));
-            zapatillaService.save(nombre, precio, color, talla, stock, idMarca, idModelo, nombreImagen);
-        } catch (Exception e) {
-            e.printStackTrace();
-            PRG.error("El Producto " + nombre + " ya existe", "/zapatilla/c");
+    public String cPost(@RequestParam("nombre") String nombre,
+                        @RequestParam("precio") double precio,
+                        @RequestParam("color") String color,
+                        @RequestParam("talla") String talla,
+                        @RequestParam("stock") Integer stock,
+                        @RequestParam("idMarca") Long idMarca,
+                        @RequestParam("idModelo") Long idModelo,
+                        @RequestParam("imagenZapatilla") MultipartFile imagenZapatilla,
+                        HttpSession session) throws DangerException {
+        Usuario usuario = (Usuario) session.getAttribute("usuario");
+        if (usuario == null || !Boolean.TRUE.equals(usuario.getAdmin())) {
+            return "redirect:/login";
         }
-        return "redirect:/admin";
-    }else{
-        return "Image Empty";
-    }
+        if (!imagenZapatilla.isEmpty()) {
+            try {
+                String imageName = imagenZapatilla.getOriginalFilename();
+                Path destPath = Paths.get(LOAD_DIR, imageName);
+                Files.copy(imagenZapatilla.getInputStream(), destPath, StandardCopyOption.REPLACE_EXISTING);
+                String nombreImagen = imageName.substring(imageName.lastIndexOf("\\") + 1, imageName.lastIndexOf("."));
+                zapatillaService.save(nombre, precio, color, talla, stock, idMarca, idModelo, nombreImagen);
+            } catch (Exception e) {
+                e.printStackTrace();
+                PRG.error("El Producto " + nombre + " ya existe", "/zapatilla/c");
+            }
+            return "redirect:/admin";
+        } else {
+            return "Image Empty";
+        }
     }
 
     @GetMapping("u")
-    public String update(
-            @RequestParam("id") Long idZapatilla,
-            ModelMap m) {
+    public String update(@RequestParam("id") Long idZapatilla, ModelMap m, HttpSession session) {
+        Usuario usuario = (Usuario) session.getAttribute("usuario");
+        if (usuario == null || !Boolean.TRUE.equals(usuario.getAdmin())) {
+            return "redirect:/login";
+        }
         m.put("zapatilla", zapatillaService.findById(idZapatilla));
         m.put("marcas", marcaService.findAll());
         m.put("modelos", modeloService.findAll());
@@ -96,17 +107,20 @@ public class ZapatillaController {
     }
 
     @PostMapping("u")
-    public String updatePost(
-            @RequestParam("idZapatilla") Long idZapatilla,
-            @RequestParam("nombre") String nombre,
-            @RequestParam("precio") double precio,
-            @RequestParam("color") String color,
-            @RequestParam("talla") String talla,
-            @RequestParam("stock") Integer stock,
-            @RequestParam("imagenZapatilla") MultipartFile imagenZapatilla,
-            @RequestParam("idMarca") Long idMarca,
-            @RequestParam("idModelo") Long idModelo
-            ) throws DangerException {
+    public String updatePost(@RequestParam("idZapatilla") Long idZapatilla,
+                             @RequestParam("nombre") String nombre,
+                             @RequestParam("precio") double precio,
+                             @RequestParam("color") String color,
+                             @RequestParam("talla") String talla,
+                             @RequestParam("stock") Integer stock,
+                             @RequestParam("imagenZapatilla") MultipartFile imagenZapatilla,
+                             @RequestParam("idMarca") Long idMarca,
+                             @RequestParam("idModelo") Long idModelo,
+                             HttpSession session) throws DangerException {
+        Usuario usuario = (Usuario) session.getAttribute("usuario");
+        if (usuario == null || !Boolean.TRUE.equals(usuario.getAdmin())) {
+            return "redirect:/login";
+        }
         try {
             String imageName = imagenZapatilla.getOriginalFilename();
             Path destPath = Paths.get(LOAD_DIR, imageName);
@@ -114,14 +128,17 @@ public class ZapatillaController {
             String nombreImagen = imageName.substring(imageName.lastIndexOf("\\") + 1, imageName.lastIndexOf("."));
             zapatillaService.update(idZapatilla, nombre, precio, color, talla, stock, nombreImagen, idMarca, idModelo);
         } catch (Exception e) {
-            PRG.error("El Producto" + nombre + " ya está registrado", "/zapatilla/r");
+            PRG.error("El Producto " + nombre + " ya está registrado", "/zapatilla/r");
         }
         return "redirect:/zapatilla/r";
     }
 
     @PostMapping("d")
-    public String delete(
-            @RequestParam("id") Long idZapatilla) throws DangerException {
+    public String delete(@RequestParam("id") Long idZapatilla, HttpSession session) throws DangerException {
+        Usuario usuario = (Usuario) session.getAttribute("usuario");
+        if (usuario == null || !Boolean.TRUE.equals(usuario.getAdmin())) {
+            return "redirect:/login";
+        }
         try {
             zapatillaService.delete(idZapatilla);
         } catch (Exception e) {
